feat: legacy claim materialize (#52)
This is a port of the code from
https://github.com/storacha/content-claims/blob/bffff9d9c7398daa8c197b2b98af2dd9eee33de3/packages/infra/src/lib/store/block-index.js#L38-L85

The possibly contentious issue is that I'm using testcontainers in the
AWS tests to spin up a real life dynamodb instance. Unfortunately
testcontainers do not work in mac/win (I didn't realise until after),
but given we don't deploy to these environments I don't think it's
really an issue. The one dependency it introduces is thta you need
docker running on your local dev system to run ALL the tests now.

TODO
* [x] update AWS lambda code to make use of the legacy claim
materialization.