chore: upgrade goreleaser to v2 (#76)
Upgrade `goreleaser/goreleaser-action` to v6, which will use goreleaser
v2, and the Go version being used. Minor adjustments to the config were
required according to [the
docs](https://goreleaser.com/deprecations/#removed-in-v2).

This should solve the error in the GoReleaser workflow. It complains
because the go version line in `go.mod` (which currently reads `go
1.23.2`) shouldn't include the patch number. However, the patch is
supported in `go.mod` since Go 1.21.