chore: change where mocks are generated (#88)
#87 required the introduction of `MockProviderIndex`, which imports the
`providerindex` package. At the same time, the tests in the
`providerindex` package itself import the `mocks` package, where all
mocks are generated currently. This causes an import cycle. More import
cycles can be created in the future as we generate mocks for other
interfaces.

There are two ways of implementing a future-proof solution for these
cycles:
- keep all mocks in the same package and move tests to their own
packages. This has the advantage that mock code lives in a single
location and doesn't clutter actual business logic code. The downside is
that non-exported members and functions cannot be tested directly as the
tests now call the code under testing as an external package. We
currently have some white box tests that call non-exported functions and
I don't think that's bad (if not mis-used).
- generate mocks in the same package the interfaces they are mocking
live in (except for external interfaces, i.e. those defined in
dependencies used by the project). This spreads mock code files across
the codebase, but allows test code to be in the same package than the
code under test. It's not completely weird that mocks are offered by the
same packages that export the relevant interfaces.

I decided to go with the second option in this case cause it allows the
most flexibility when writing tests. That implied changing mock
generation configuration, moving mocks around and updating the imports
in test files using them.