feat: add support for OpenTelemetry (#96)
Ref: https://github.com/storacha/project-tracking/issues/246

This PR adds support for OpenTelemetry instrumentation.

The traces are quite basic for now, with a single span for the lambda
handler that is invoked, which gives little info. I'm following up with
another PR that adds more details about outgoing requests to other
services.

## Implementation details

The main issue I faced with this was related with lambda lifecycles. It
seems that a lambda can be frozen at any given time once the handler has
been executed. OpenTelemetry defaults to using batching and asynchronous
processing of traces and spans, which doesn't play well with the lambda
being sent to sleep potentially at any moment.

It is indeed possible to export traces directly from the handler logic,
but it needs to be done synchronously. As this would have an impact on
request processing times, the recommended practice for production
deployments is to offload trace batching and exporting to an independent
collector process.

This is what the [AWS Distro for OpenTelemetry Lambda
layer](https://aws-otel.github.io/docs/getting-started/lambda) offers.
It is a lite version of the OpenTelemetry collector that exports
telemetry data to X-Ray by default, but can be configured to export to
other backends.

Configuration is a bit cumbersome because a configuration file is
required, that's why I added that configuration file and changed the
Makefile and the terraform script so that it gets bundled with the
lambda binary.