fix: provider store atomic add to set (#99)
This PR alters the provider store so that adding a `ProviderResult` to
the set for a given hash is an atomic operation.

It uses Redis `SAdd` (add to set) and `SMembers` (get members of a set).

Previously we'd read the current value, append to the list and then
write the value back. I was seeing intermittent test failures in
https://github.com/storacha/testthenetwork - one of the index provider
result and the location provider result were being dropped because they
were being added to the set at approximately the same time.