fix(tests): flaky test in BucketFallbackMapper tests (#104)
Fixes #103 

The original implementation launched an HTTP server for the test
concurrently in a different goroutine. This poses a race condition
between starting the server and running the tests. Sometimes, the test
was run before the server was ready and the test failed.

Finding the cause has been tricky because, in case of an error in the
HEAD request, `BucketFallbackMapper` swallows the original error and
returns `ErrKeyNotFound`, masking the real cause of the issue. This
error handling has been improved to avoid this.

To fix the test, the regular HTTP server being used has been replaced
with an `httptest.Server`. It simplifies the test code as handling its
shutdown is easier, but the reason it fixes the issue is that
`httptest.Server` prepares the network listener synchronously, and then
starts the server concurrently. That means that, even in the case the
test is run before the server is ready to server requests, the listener
will be able to buffer them and prevent the test from failing.