refactor(providerindex): accept multiple legacy options (#86)
# Goals

Get a full roundtrip to work with freeway -- these are all the
additional changes required

# Implementation

- update go-capabilities to include fixes from
https://github.com/storacha/go-capabilities/pull/6 (which had not been
updated in a minute, so there are a few changes required)
- modify the legacy claims finder to work with multiple claim mappers,
and be sensitive to target claims. the reason for this is as follows --
I had assumed by fallback mapper could simply run on top of the base
legacy mapper and run when the legacy mapper returns no claims. however,
in practice, we end up in a situation where we are specifically looking
for a Shard CID location claim, and the base legacy mapper returns a
Shard CID EQUALS claim. So we need to have that filtered out and then
run my fallback mapper anyway. This ends up making much more sense to do
at the legacy claims finder level than in the fallback mapper.

# For Discussion

Still need to add some new test cases for the legacy claims finders now
that it uses multiple mappers & handles filtering

---------

Co-authored-by: Vicente Olmedo <vicente@storacha.network>