feat: allow specifying query type in HTTP API (#100)
💯 

closes #94

This PR adds a new `kind` query parameter to the `/claims` endpoint that
indicates which claim types should be returned. It has three possible
values:
- `standard`: this is the default, all three types of claims (location,
index and equals) are returned
- `location`: only location claims are returned
- `index_or_location`: location and index claims are returned

Aside from adding support in both the server and the client, a new
`type`/`t` parameter is added to the `indexer` CLI so that different
query types can be performed.