fix: distributed tracing not working (#139)
fixes #134 

Context propagation was not properly configured and incoming trace
context was not being extracted from requests coming from the gateway.

Interestingly, telemetry was already configured to sample only requests
that were sampled in the upstream service, but since context propagation
was not working the result was that all requests were being sampled.
Fixing context propagation also means that the indexer will only sample
requests that are sampled in the gateway.