fix: move block index table legacy service to regular legacy claims flow (#138)
Resolves #135 

This PR takes the logic from the block index table legacy service and
turns it into a ContentToClaimsMapper. Then adds that mapper to the list
of mappers used by the regular legacy claims flow so that its results
are handled and cached the same way as with the results from other
mappers.

---------

Co-authored-by: Hannah Howard <hannah@hannahhoward.net>