fix(sqs): adjust visibility timeout to lambda timeout (#167)
We increased the providercache lambda timeout to 300s but didn't update
the visibility timeout in the SQS queue it is consuming messages from.
That makes the messages in the batch being processed available to other
lambdas before the one processing them has the chance to delete them
and/or mark them as processed.