feat(infra): self-defined cache cluster with LFU eviction policy (#163)
Change our current valkey serverless cluster into a valkey cluster with
cluster mode enabled and update service redis clients to the new
architecture.

## Considerations
### Cluster mode
There are two types of Elasticache clusters: with cluster mode disabled
(CMD) and with cluster mode enabled (CME). CMD clusters are essentially
a primary node with some optional replicas for high availability. Only
one of these node groups is possible with CMD. CME, of course, allows
several node groups (i.e. shards) and will do sharding of data slots
between groups.

In our use case, since we are only defining a single shard, both are
similar. The main difference is that, when cluster mode is enabled, the
cluster offers a configuration endpoint that can be used by the client
to automatically discover nodes to route requests to. The `go-redis`
library we use handles transparently what node to send requests to when
in cluster mode. With CMD you get a primary endpoint and a reader
endpoint and we'd need to decide ourselves where to send each type of
query.

I went with CME because it makes client logic a bit more
straightforward, but using CMD shouldn't be difficult. I'll keep an eye
on the metrics to see if using CME comes with a performance penalty.

### Forcing deployments to use a Redis/Valkey cluster
I changed all redis clients to ClusterClients, which forces standalone
deployments to use a cluster too. This is not ideal if we want to see
users run their own indexers, but the reality is that nobody is doing
that today. It can probably be solved by using `redis.UniversalClient`,
but we can do it in the future when this is an actual problem.

## Tests
I did some manual testing and the traces confirmed that using `ReadOnly:
true` and `RouteRandomly: true` makes the service send requests to the
three nodes in the cluster.

## Note for reviewers
This PR is based on my previous PR for multi-region deployments #154.
I'll rebase on top of main once the other one is merged.