perf: concurrent provider cache (#173)
This PR alters the provider cacher code to concurrently cache provider
records up to 5 at a time.

In manual testing with an atrificial network delay I was able to see
caching 10k provider records reduce from 20s to around 5s.

The aim is to process the caching queue faster as it is continually
growing atm.