fix(ci): run tofu plans for prod when not releasing (#181)
We currently only run plans of our terraform config for staging. Our
config for prod is slightly different from that of staging, and today we
only catch issues when we apply the config to prod as part of a release.

This PR updates our CI so that plans are run for prod. That way we can
anticipate and fix issues when updating the config before they get
merged.