chore(dns): alternate domain for API gateway, main domain for cloudfront (#180)
For prod, we want the main `indexer.storacha.network` domain to point to
the cloudfront distribution. The current config uses that domain name
both as a custom domain for API gateway and as an alternate domain for
the cloudfront distribution.

Instead of that, we will give the API gateway its own
`api.indexer.storacha.network` domain to make it clearer which resource
is using which domain name. This allows removing references to the
gateway's custom domain in cloudfront route53 records, and references to
the cloudfront domain in the gateway's config.