feat: configure IPNI endpoint and announce URLs (#205)
Allows the IPNI endpoint for queries to be configured so that when we
have deployed our own IPNI instance we can use it for querying.

Allows the IPNI announce URLs to be configured so that when we have
deployed our own IPNI instance we can announce to it as well as
cid.contact.