feat: redis pipelining (#206)
This PR adds redis pipelining to the `ProviderIndex` cache and uses it
whenever we need to add a batch of data in one go.

We upped the concurrency count for writes from 5 to 500, and while this
made a difference, it was still not able to write 10k items within 30s.
Pipelining removes round trips and hopefully allows data to be pumped
into redis as fast as the network connection between the lambda and the
cache will allow or as fast as redis (valkey) can write to memory.

Note:

* I removed the `written` count return value for
`CacheProviderForIndexRecords` as this is harder to implement with
pipeline and not actually used in production code.