fix: limit results from legacy block index table (#222)
Some blocks are in MANY MANY CARs - sometimes because it's a popular
block (`LICENCE.md` for example) and other times we had a bug that meant
that the same content got uploaded many times over in different CAR
files.

This PR adds a limit of 25 to the legacy block index table query
results. The same value that is used in the legacy content claims
service.

It prevents the indexing service from using too much cache space for a
given hash and ensures quick response times in cases where a block is
found in many CARs.