feat: switch to a storoku-based deployment (#230)
closes https://github.com/storacha/indexing-service/issues/224

This PR removes the existing deployment config and uses storoku to set
up a new deployment strategy, moving off from the lambdas to ECS tasks.

The config based in environment variables is split now in two different
places. On the one hand, storoku deployment config already passes many
env vars related with infra resources. The remaining environment
variables are defined in
[deploy/.env.production.local.tpl](https://github.com/storacha/indexing-service/compare/vic/feat/containerize-for-aws...vic/feat/switch-to-storoku?expand=1#diff-07487185485873d9f33d651e0d9868364c1db2703df6481bbd5f2770c856e93b).

Some of the configuration in the existing deployment was not supported
by storoku, so I had to add some new features:
- https://github.com/storacha/storoku/pull/7
- https://github.com/storacha/storoku/pull/8
- https://github.com/storacha/storoku/pull/9
- https://github.com/storacha/storoku/pull/12
- https://github.com/storacha/storoku/pull/15
- https://github.com/storacha/storoku/pull/16
- https://github.com/storacha/storoku/pull/17
- https://github.com/storacha/storoku/pull/19

The PR is large for my liking, but most of it are the old terraform
files being deleted and the new ones from storoku being added.