feat(server): add url to mimic IPNI server (#246)
# Goals

IPNI has agreed to query Storacha directly from cid.contact when it
receives queries. However, it needs an endpoint that speaks the IPNI API
format. This PR turns our indexer into a kind of IPNI instance -- note
one that simply returns back our HTTP gateway endpoint for retrievals.
(similar to our router v1)

# Implementation

- add optional `/cid/{cid}` route to the server
- in this route, run the normal query, and simply return a hardcoded
result in IPNI response format if there is 1 or more location claims for
the CID
   - otherwise return 404
- the hardcoded peer address & metadata are passed into configuration
for the server
- add mechanisms to turn on IPNI route to both manual server and AWS
server
- add AWS configs for turning on the IPNI route
- add vars to github deploy and terraform env to use the IPNI endpoint
- add a var that can be turned on to point to prod resources from a
non-prod environment (needed this for testing)

# For Discussion

I pushed all this out to my personal dev env and it can be tested via:
`https://hannah.indexer.storacha.network/cid/{any valid Storacha cid}`

I did consider that we could get into some kind of infinite loop since
we also hit IPNI on the backend, but since we use a custom IPNI instance
that doesn't do this fallback we should be fine.

Obviously all this is a bit stopgap and I hope we'll get to turn it off
someday soon.