refactor: expect attached retrieval authorization in `assert/index` invocation (#263)
Instead of obtaining a service retrieval authorization, expect the
`assert/index` invocation to include a delegation for retrieving the
index blob.

This allows clients to invoke `assert/index` directly, not as a proxy
through `space/index/add`.

The delegation is expected to be referenced from the `assert/index`
invocation `facts` under the key "auth".