feat: send traces to Honeycomb or Storacha's OTEL collector (#277)
Adds a new `TELEMETRY_DISABLED` environment variable that can be used to
enable/disable telemetry, instead of relying on the `HONEYCOMB_API_KEY`
variable. When telemetry is enabled, traces will be sent to Honeycomb if
`HONEYCOMB_API_KEY` is not empty, and to our own OTEL collector if it
isn't.

`prod` and `staging` will keep sending traces to Honeycomb, while
`warm-staging` and `forge-prod` will send them to our OTEL collector so
they end up in Grafana.