fix: use built-in gzip support in stdlib's HTTP client  (#281)
Go's HTTP client in the standard library supports gzip automatically
when one doesn't set the `Accept-Encoding: gzip` header. If the header
is not set, the client will do it automatically (unless it's disabled in
the underlying transport) and will also handle compression/decompression
internally.

I removed the code setting the `Accept-Encoding` header and also manual
comp/decomp and the implementation still passes the tests.